Add react-router and extract App into GameView
