+++ package.json
@@ -21,6 +21,7 @@
     "react": "^19.1.0",
     "react-confetti": "^6.4.0",
     "react-dom": "^19.1.0",
+    "react-router": "^7.6.2",
     "react-use": "^17.6.0",
     "use-sound": "^5.0.0",
     "vite-express": "^0.21.1"
